Downgrade history version to 4.7.2 to fix warning issue Refactor history services
