+++ example/src/components/index.js
@@ -1,3 +1,2 @@
 export { default as SideBar } from './SideBar';
-export { default as Navigation } from './Navigation';
 export { default as CustomLoading } from './CustomLoading';
