+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-router-guard",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "description": "React router guard",
   "author": "laptransang",
   "license": "MIT",
@@ -64,14 +64,13 @@
     "dist"
   ],
   "dependencies": {
-    "history": "^4.9.0",
+    "history": "4.7.2",
     "react-loadable": "^5.5.0",
     "react-router-dom": "^4.3.1"
   },
   "keywords": [
     "react router guard",
-    "react authentication route",
-    "how to implement authenticated routes",
-    "private route"
+    "react authentication router",
+    "react router code splitting"
   ]
 }
