+++ src/components/AsyncComponent/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { promiseSerial, verifyRouterData, makeCancelable } from 'utils/helpers';
-import loadingService from 'services/loadingService';
+import { loadingService } from 'services';
 
 const propTypes = {
   promise: PropTypes.oneOfType([PropTypes.array, PropTypes.func]).isRequired,
