--- src/history.js
@@ -1,5 +1,0 @@
-import createBrowserHistory from 'history/createBrowserHistory';
-
-const history = createBrowserHistory();
-
-export default history;
