+++ src/index.js
@@ -1,4 +1,4 @@
 export { default as dynamicWrapper } from 'services/dynamicWrapper';
 export { default as RouterGuard } from 'RouterGuard';
 export { Link, Route, Redirect, Router, BrowserRouter, Switch, withRouter, NavLink } from 'react-router-dom';
-export { default as history } from 'history';
+export { default as history } from 'services/history';
