+++ src/services/index.js
@@ -1,0 +1,4 @@
+export { default as dynamicWrapper } from 'dynamicWrapper';
+export { default as history } from 'history';
+export { default as loadingService } from 'loadingService';
+export { default as routes } from 'routes';
